--- Assets/LoadingScreen.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0c64e527f2baee946a22c1cc9182b05d
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
