--- Assets/SceneMan.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9aeab9423fa9be748a15569e6d78c927
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
