--- Assets/Scenes/Map.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 079492ba1bea71240aa6ac18cc1087b3
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
