--- Assets/Script/MainMenu.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class MainMenu : MonoBehaviour
-{
-    public void PlayGame()
-    {
-        SceneManager.LoadScene("Map");
-        Debug.Log("play");
-    }
-
-    public void ExitGame()
-    {
-        Application.Quit();
-        Debug.Log("shoo shoo");
-    }
-}
