--- Assets/Script/MainMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eff58d4f39538d242ab8f66af3d50b7f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
