+++ ProjectSettings/EditorBuildSettings.asset
@@ -15,6 +15,6 @@ EditorBuildSettings:
     path: Assets/Scenes/Scene1.unity
     guid: c1785d931b8c9574fbee2ae7b613e9c1
   - enabled: 1
-    path: Assets/Scenes/Map.unity
+    path: Assets/Scenes/Game.unity
     guid: 079492ba1bea71240aa6ac18cc1087b3
   m_configObjects: {}
